feat: Adding data generator utility functions for integration tests and updating tests
