+++ server/__tests__/routeTests/league.routes.test.js
@@ -2,6 +2,7 @@
 
 window.setImmediate = window.setTimeout
 
+const TestDataGenerator = require("../../utilityFunctions/testDataGenerator.js");
 const request = require("supertest");
 const express = require("express");
 const LeagueRoutes = require("../../routes/league.routes");
@@ -12,18 +13,7 @@ app.use("/league", LeagueRoutes);
 app.use(Middlewares.errorHandler);
 const testDb = require("../../models");
 
-const createDummyGroup = async function(groupName){
-  let newGroupData = {
-    name: groupName,
-    street_address: "123 Main Street",
-    city: "Sample City",
-    state: "CA",
-    zip_code: "12345",
-    logo_url: "https://example.com/logo",
-  }
-  const sampleGroup = await testDb.Group.create(newGroupData);
-  return sampleGroup;
-}
+
 
 describe("League Routes", () => {
 
@@ -35,7 +25,7 @@ describe("League Routes", () => {
   // ------------------- Get League by Group ID Tests ----------------
 
   it("should handle GET /getLeagueByGroupId", async () => {
-    const groupM = await createDummyGroup("Group Uno");
+    const groupM = await TestDataGenerator.createDummyGroup("Group Uno");
 
     await testDb.League.create({ group_id: groupM.id, name: "Leeroy League" });
     await testDb.League.create({ group_id: groupM.id, name: "Martin Martians" });
@@ -49,7 +39,7 @@ describe("League Routes", () => {
   });
 
   it("should not get any leagues with GET /getLeagueByGroupId", async () => {
-    const groupM = await createDummyGroup("Group Uno");
+    const groupM = await TestDataGenerator.createDummyGroup("Group Uno");
 
     const response = await request(app)
       .get(`/league/${groupM.id}`)
@@ -64,7 +54,7 @@ describe("League Routes", () => {
   // ------------------- Create Tests ----------------
 
   it("should handle POST /create", async () => {
-    const groupM = await createDummyGroup("Salinas");
+    const groupM = await TestDataGenerator.createDummyGroup("Salinas");
 
     const response = await request(app)
       .post("/league/")
@@ -78,7 +68,7 @@ describe("League Routes", () => {
   });
 
   it("should not create if name is not unique POST /create", async () => {
-    const groupM = await createDummyGroup("Saul's Group");
+    const groupM = await TestDataGenerator.createDummyGroup("Saul's Group");
 
     await testDb.League.create({ group_id: groupM.id, name: "Salinas" });
 
@@ -93,8 +83,8 @@ describe("League Routes", () => {
   });
 
   it("should create a league with the same name from a different group POST /create", async () => {
-    const groupUno = await createDummyGroup("Watsonville Corp.");
-    const groupDos = await createDummyGroup("Salinas Inc.");
+    const groupUno = await TestDataGenerator.createDummyGroup("Watsonville Corp.");
+    const groupDos = await TestDataGenerator.createDummyGroup("Salinas Inc.");
 
     await testDb.League.create({ group_id: groupUno.id, name: "Summer 2024" });
 
@@ -112,7 +102,7 @@ describe("League Routes", () => {
   // ------------------- Update Tests ----------------
 
   it("should update league with valid ID and input PATCH /patch", async () => {
-    const groupM = await createDummyGroup("Salinas");
+    const groupM = await TestDataGenerator.createDummyGroup("Salinas");
     const league = await testDb.League.create({ group_id: groupM.id, name: "SOMOS" });
 
     const updatedLeagueName = "Sopa Marucha";
@@ -127,7 +117,7 @@ describe("League Routes", () => {
   });
 
   it("should return an error if league not found PATCH /patch", async () => {
-    const nonExistentLeagueId = 9999; 
+    const nonExistentLeagueId = "9999"; 
 
     const response = await request(app)
       .patch(`/league/${nonExistentLeagueId}`)
@@ -140,7 +130,7 @@ describe("League Routes", () => {
   });
 
   it("should not update if the new name is already used in the group", async () => {
-    const groupM = await createDummyGroup("Salinas");
+    const groupM = await TestDataGenerator.createDummyGroup("Salinas");
   
     const league1 = await testDb.League.create({ group_id: groupM.id, name: "Shrek League" });
     const league2 = await testDb.League.create({ group_id: groupM.id, name: "Donkey League" });
@@ -161,7 +151,7 @@ describe("League Routes", () => {
   // ------------------- Delete Tests ----------------
 
   it("should delete a league with a valid league ID DELETE /delete", async () => {
-    const groupM = await createDummyGroup("Salinas");
+    const groupM = await TestDataGenerator.createDummyGroup("Salinas");
     const league = await testDb.League.create({ group_id: groupM.id, name: "SOMOS", });
 
     const response = await request(app)
